Implement Account methods: list_tx(), utxos(), balance(), create_tx() and sign()
Refactored out of WalletCtx
